renamed files and setup highlights component
